1 Fix the build error caused by the modification for compiler 6. 2 Revise startup_M058.s and retarget.c files to fix build issue. 3 Modify settings of Keil project to speed up compilation. 4 Revise copyright format for gen. chm file. 5 Fix download issue of IAR Project.

+++ SampleCode/StdDriver/I2C_Wakeup_Slave/main.c
@@ -7,8 +7,9 @@
  *           Show how to wake up MCU from Power-down mode through I2C interface.
  *           This sample code needs to work with I2C_Wakeup_Master.
  * @note
- * Copyright (C) 2014 Nuvoton Technology Corp. All rights reserved.
+ * @copyright SPDX-License-Identifier: Apache-2.0
  *
+ * @copyright Copyright (C) 2014 Nuvoton Technology Corp. All rights reserved.
  ******************************************************************************/
 #include <stdio.h>
 #include "M058S.h"
@@ -323,7 +324,7 @@ int32_t main(void)
     }
 
     /* Wake-up Interrupt Message */
-    printf("Power-down Wake-up INT 0x%x\n", ((CLK->PWRCON) & CLK_PWRCON_PD_WU_STS_Msk));
+    printf("Power-down Wake-up INT 0x%x\n",(unsigned int)((CLK->PWRCON) & CLK_PWRCON_PD_WU_STS_Msk));
     printf("I2C0 WAKE INT 0x%x\n", I2C0->I2CWKUPSTS);
 
     /* Disable power wake-up interrupt */
